requirements.txt
bokeh==2.4.3
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
Pillow==10.2.0
streamlit==1.29.0
streamlit_extras==0.3.6
sympy==1.12
